+++ src/fsql/fsql-grammar-utils.ts
@@ -43,9 +43,9 @@ export namespace FsqlGrammarUtils {
     return parentNode
   }
 
-  export function getReferencedTypes(
+  export function getTypes(
     parsingResult: any,
-  ): { typeName: moo.Token; as: moo.Token | undefined }[] {
+  ): { typeName: moo.Token; as: moo.Token | undefined; path: string }[] {
     const types = traverse(parsingResult).reduce(function(acc, v) {
       if (this.key !== 'typeName') {
         return acc
@@ -60,9 +60,74 @@ export namespace FsqlGrammarUtils {
         'as' in parentNode &&
         parentNode['as']
       ) {
-        acc.push({ typeName: v, as: parentNode['as'] })
+        acc.push({
+          typeName: v,
+          as: parentNode['as'],
+          path: this.path.join(','),
+        })
       } else {
-        acc.push({ typeName: v, as: undefined })
+        acc.push({ typeName: v, as: undefined, path: this.path.join(',') })
+      }
+
+      return acc
+    }, [])
+
+    return types
+  }
+
+  export function getSubqueries(
+    parsingResult: any,
+  ): { subquery: moo.Token; as: moo.Token | undefined; path: string }[] {
+    const types = traverse(parsingResult).reduce(function(acc, v) {
+      if (this.key !== 'subquery') {
+        return acc
+      }
+      if (v === undefined || v === null) {
+        return acc
+      }
+
+      const parentNode = this.parent!.node
+      if (
+        parentNode.type === 'table_expression' &&
+        'as' in parentNode &&
+        parentNode['as']
+      ) {
+        acc.push({
+          subquery: v,
+          as: parentNode['as'],
+          path: this.path.join(','),
+        })
+      } else {
+        acc.push({ subquery: v, as: undefined, path: this.path.join(',') })
+      }
+
+      return acc
+    }, [])
+
+    return types
+  }
+
+  export function getSelects(
+    parsingResult: any,
+  ): { term: any; as: moo.Token | undefined; path: string }[] {
+    const types = traverse(parsingResult).reduce(function(acc, v) {
+      if (this.key !== 'term') {
+        return acc
+      }
+      if (v === undefined || v === null) {
+        return acc
+      }
+
+      // This will skip all `*` expression
+      const parentNode = this.parent!.node
+      if (
+        parentNode.type === 'select_expression' &&
+        'as' in parentNode &&
+        parentNode['as']
+      ) {
+        acc.push({ term: v, as: parentNode['as'], path: this.path.join(',') })
+      } else {
+        acc.push({ term: v, as: undefined, path: this.path.join(',') })
       }
 
       return acc