+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Add a fsql to raw sql command `Hybris: Translate to Raw SQL`
 - Add a databaseColumn in the definition provider for types
+- Support comments in fsql
+- Add some snippets for autocompletion i.e. `WHERE`, `JOIN`, `ON`, `AS`, `AND`, `OR`
 
 ## 0.0.6